+++ examples/high_level/complete_graph_tdvp.jl
@@ -5,7 +5,7 @@ limits = Limits(
     maxdim = 100,
 )
 
-output(n) =
+measurements(n) =
     "data" => [X, Y, Z, Y(1)Y(2), Z(1)Y(2)Y(3), EntanglementEntropy(n ÷ 2, 4), Purity, Trace]
 
 sim_data(n) = SimData(
@@ -17,7 +17,7 @@ sim_data(n) = SimData(
     phases = [
         create_graph_state(complete_graph(n); limits),
         ToMixed(
-            final_measures = output(n),
+            final_measures = measurements(n),
             limits = limits,
         ),
         Evolve(
@@ -26,7 +26,7 @@ sim_data(n) = SimData(
             time_step = 0.025,
             algo = Tdvp(),
             evolver = sum(Dissipator(Sp)(i) for i in 1:n),
-            measures = output(n),
+            measures = measurements(n),
             measures_period = 4,
         ),
     ],